+++ .github/templates/macos_binary_build_workflow.yml.j2
@@ -111,6 +111,8 @@ jobs:
 
           # shellcheck disable=SC2086
           python "${PYTORCH_ROOT}/.ci/pytorch/smoke_test/smoke_test.py" --package torchonly ${SMOKE_TEST_PARAMS}
+
+          python "${PYTORCH_ROOT}/.ci/pytorch/check_wheel_tags.py"
 {%- endif %}
       - uses: actions/upload-artifact@v4.4.0
         if: always()
